:lipstick: Rename main activities
