+++ exampleApp2/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <application
         android:icon="@drawable/ic_udinic"
         android:label="@string/app_name">
-        <activity android:name=".Main2">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <category android:name="android.intent.category.LAUNCHER" />
                 <action android:name="android.intent.action.MAIN" />
