+++ exampleApp2/src/main/res/layout/activity_main.xml
